improved-agent-erc3/requirements.txt
 --extra-index-url https://erc.timetoact-group.at/
erc3>=1.0.8
openai>=2.8.1
google-generativeai>=0.8.0
